Moved AspectProxy to HttpGetRequest
